Use common implementation for R package detection
