+++ apps/lsp/tsconfig.json
@@ -3,9 +3,11 @@
     "lib": ["ES2020"],
     "module": "CommonJS",
     "outDir": "./dist",
-    "rootDir": "./src",
     "sourceMap": true,
     "resolveJsonModule": true,
+    "paths": {
+      "@utils/*": ["../utils/*"]
+    }
   },
   "exclude": ["node_modules"],
   "extends": "tsconfig/base.json",
