+++ apps/vscode/src/r-utils.ts
@@ -1,0 +1,29 @@
+/*
+ * r-utils.ts
+ *
+ * Copyright (C) 2025 by Posit Software, PBC
+ *
+ * Unless you have received this program directly from Posit Software pursuant
+ * to the terms of a commercial license agreement with Posit Software, then
+ * this program is licensed to you under the terms of version 3 of the
+ * GNU Affero General Public License. This program is distributed WITHOUT
+ * ANY EXPRESS OR IMPLIED WARRANTY, INCLUDING THOSE OF NON-INFRINGEMENT,
+ * MERCHANTABILITY OR FITNESS FOR A PARTICULAR PURPOSE. Please refer to the
+ * AGPL (http://www.gnu.org/licenses/agpl-3.0.txt) for more details.
+ *
+ */
+
+import * as vscode from 'vscode';
+import { isRPackage as isRPackageImpl } from "@utils/r-utils";
+
+// Version that selects workspace folder
+export async function isRPackage(): Promise<boolean> {
+  if (vscode.workspace.workspaceFolders === undefined) {
+    return false;
+  }
+
+  // Pick first workspace
+  const workspaceFolder = vscode.workspace.workspaceFolders[0];
+
+  return isRPackageImpl(workspaceFolder.uri);
+}
