+++ apps/vscode/tsconfig.json
@@ -5,7 +5,11 @@
     "outDir": "out",
     "lib": ["ES2021"],
     "sourceMap": true,
-    "strict": true /* enable all strict type-checking options */
+    "strict": true,
+    "paths": {
+      "@utils/*": ["../utils/*"]
+    }
+    /* enable all strict type-checking options */
     /* Additional Checks */
     // "noImplicitReturns": true, /* Report error when not all code paths in function return a value. */
     // "noFallthroughCasesInSwitch": true, /* Report errors for fallthrough cases in switch statement. */
